+++ app/actions/logs.ts
@@ -4,7 +4,7 @@ import { logs, endpoints } from '@/lib/db/schema'
 import { eq, desc } from 'drizzle-orm'
 import { revalidatePath } from 'next/cache'
 import { db } from '@/lib/db'
-import { authenticatedAction } from '@/lib/data/safe-action'
+import { authenticatedAction } from '@/lib/data/safe-action.server'
 import { z } from 'zod'
 
 /**
