+++ lib/auth/index.ts
@@ -1,7 +1,7 @@
 import 'server-only';
 import NextAuth from "next-auth";
 import { DrizzleAdapter } from "@auth/drizzle-adapter";
-import { db } from "../db";
+import { db } from "@/lib/db";
 import type { NextAuthConfig } from "next-auth";
 import { User } from "next-auth";
 import Resend from "next-auth/providers/resend";
